// web/src/generated/models/skillDefinitionCreated.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Priv-View Core API
 * Phase-one Core API for authentication, hierarchical tenants, clients, matters, and matter metadata definitions.
 * OpenAPI spec version: 0.1.0
 */
import type { SkillDefinitionRead } from './skillDefinitionRead';
import type { SkillDefinitionVersionRead } from './skillDefinitionVersionRead';

export interface SkillDefinitionCreated {
  skill: SkillDefinitionRead;
  version: SkillDefinitionVersionRead;
}