// web/src/generated/models/workflowSpecRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Priv-View Core API
 * Phase-one Core API for authentication, hierarchical tenants, clients, matters, and matter metadata definitions.
 * OpenAPI spec version: 0.1.0
 */
import type { WorkflowRoleSpecRead } from './workflowRoleSpecRead';

export interface WorkflowSpecRead {
  key: string;
  code_version: string;
  name: string;
  description: string;
  roles: WorkflowRoleSpecRead[];
}